+++ src/Breadcrumbs/DataAttr.js
@@ -1,0 +1,8 @@
+export const DATA_HOOKS = {
+  ITEM_WRAPPER: 'item_wrapper',
+  BREADCRUMB_CLICKABLE: 'breadcrumb-clickable',
+  BREADCRUMBS_ITEM: 'breadcrumbs-item',
+};
+
+export const DATA_ACTIVE = 'data-active';
+export const DATA_POSITION_ID = 'data-position-id';