+++ src/MultiSelect/MultiSelect.uni.driver.js
@@ -34,6 +34,8 @@ export const multiselectUniDriverFactory = base => {
       tagPrivateUniDriverFactory(
         tags.filter(async tag => (await tag._prop('id')) === tagId).get(0),
       ),
+    customSuffixExists: async () =>
+      (await getInputWrapper()).querySelector('[data-hook="custom-suffix"]')
   };
 
   return { driver: multiSelectDriver, inputDriver, dropdownLayoutDriver };