+++ src/MultiSelect/docs/ExampleReorderable.js
@@ -4,6 +4,7 @@ import React from 'react';
 import MultiSelect from 'wix-style-react/MultiSelect';
 import Card from 'wix-style-react/Card';
 import FormField from 'wix-style-react/FormField';
+import TextButton from '../../TextButton/TextButton';
 
 const options = [
   { id: '1', name: 'One', value: 'One' },
@@ -63,6 +64,7 @@ class ExampleReorderable extends React.Component {
         value={this.state.inputValue}
         onChange={this.handleOnChange}
         options={options}
+        customSuffix={<TextButton>+ Add Tag</TextButton>}
         upgrade
       />
     );