+++ src/beta/PopoverMenu/PopoverMenu.uni.driver.d.ts
@@ -3,7 +3,7 @@ import { UniDriver } from '@unidriver/core';
 export interface PopoverMenuDriver {
   exists(): Promise<boolean>;
   getTriggerElement(): UniDriver;
-  clickAtChild(): Promise<void>;
+  clickAtChild: (index: number) => Promise<void>;
   isMenuOpen(): Promise<boolean>;
   childrenCount(): Promise<number>;
 }