+++ src/index.js
@@ -136,3 +136,4 @@ export { default as Swatches } from './Swatches';
 export { default as SidebarSectionTitle } from './SidebarSectionTitle';
 export { default as SidebarSectionItem } from './SidebarSectionItem';
 export { default as Collapse } from './Collapse';
+export { default as SidebarSectionDivider } from './SidebarSectionDivider';