+++ CHANGELOG.md
@@ -15,8 +15,28 @@ Types of changes:
 
 ## Next
 
+## 6.41.0 - 2019-07-20
+
+### Fixed
+
+- `<Breadcrumbs/>` - fix UniDriver [#3777](https://github.com/wix/wix-style-react/pull/3777)
+
+- `<CardGalleryItem/>` - UniDriver - fix click function [#3744](https://github.com/wix/wix-style-react/pull/3774)
+
+- `<InputWithOptions/>` - fix Input cursor jumping to input start on ArrowUp [#3762](https://github.com/wix/wix-style-react/pull/3762)
+- `<PopoverMenu/>`, `<Swatches/>` - fix stylable colors with opacity imports [#3778](https://github.com/wix/wix-style-react/pull/3778)
+- `<Popovermenu/>` - beta - fixing testkit definition for clickAtChild [#3779](https://github.com/wix/wix-style-react/pull/3779)
+
+### Changed
+
+- `<Input/>` - replace old puppeteer driver with unidriver for testkit export [#3769](https://github.com/wix/wix-style-react/pull/3769)
+
+## 6.40.0 - 2019-07-17
+
 ### Added
 - `<SidebarSectionItem/>` - create the component (WIP) [#3751](https://github.com/wix/wix-style-react/pull/3751)
+- `<SidebarSectionDivider/>` - create the component (WIP) [#3758](https://github.com/wix/wix-style-react/pull/3758)
+- `<Multiselect/>` - add `customSuffix` prop [#3759](https://github.com/wix/wix-style-react/pull/3759)
 
 ### Fixed
 - `beta/<PopoverMenu/>` - fix `moveBy` prop for [#3755](https://github.com/wix/wix-style-react/pull/3755)